Rename directory: cmd -> gincmd
Package name was already gincmd. Changing the path to match.
