+++ gincmd/deletecmd.go
