+++ gincmd/getcontentcmd.go
