+++ gincmd/removecontentcmd.go
